添加部分注释

+++ Assets/Scripts/Player.cs
@@ -6,9 +6,9 @@ public class Player : MonoBehaviour
 {
     public float damage = 10f; 
     public float speed = 3f;                                             //玩家移动速度
-    public float HP = 30f;
-    public float HPMax = 30f;
-    public float luck = 2f;
+    public int HP = 30;
+    public int HPMax = 30;
+    public int luck = 2;
     public GameObject weapon;
 
     //玩家状态 普通状态 移动视角状态 受到攻击状态
